Makefile and testing files have been added
Makefile will compile all of the test programs. It will be up to the user to properly execute each file. 
To run the test files, simply type the following command into the terminal:

./test#

Where # refers to the digit following the name test. For example to run test0, the user would type ./test0
To run test1 the user would type ./test1
And so on